add a Docker dojo

+++ dojo.sh
@@ -3,7 +3,7 @@ set -e
 rm -rf ~/code/Dojo
 source ~/bin/list_input.sh
 
-templates=( 'Vanilla' 'React.js')
+templates=( 'Vanilla' 'React.js' 'Docker')
 list_input "Choose a template:" templates selected_template
  
 if [ "$selected_template" == "Vanilla" ]; then
@@ -18,6 +18,15 @@ if [ "$selected_template" == "React.js" ]; then
       git clone https://github.com/VKTRN/Source.git ~/code/Dojo &&
       cd ~/code/Dojo &&
       git checkout React.js
+      rm -rf .gitignore .git
+      npm install &&
+      code .
+fi
+
+if [ "$selected_template" == "Docker" ]; then
+      git clone https://github.com/VKTRN/Source.git ~/code/Dojo &&
+      cd ~/code/Dojo &&
+      git checkout Docker
       rm -rf .gitignore .git
       npm install &&
       code .
